refactor(CommandParser): CommandParser 로직 CommandType으로 이동 및 클래스 삭제
